requirements.txt
aenum==2.2.3
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
discord.py==1.3.4
idna==2.10
multidict==4.7.6
psycopg2==2.8.5
SQLAlchemy==1.3.18
setuptools==49.2.0
websockets==8.1
yarl==1.4.2

# TODO: add test libraries